colpali-api/pyproject.toml
[tool.poetry]
name = "colpali-api"
version = "0.1.0"
description = "A FastAPI application for Colpali and Vespa"
authors = ["kefio <alessandro.kayed@gmail.com>, mrpsycox <simone.silv@yahoo.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "0.111.0"
torch = "2.3.1"
colpali-engine = "0.3.5"
uvicorn = "^0.30.0"
pyvespa = "^0.51.0"
google-cloud-aiplatform = "^1.74.0"
together = "^1.3.10"
ecs-logging = "2.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.7"
pre-commit = "^3.7.1"
locust = "^2.28.0"
mypy = "1.11.0"
pytest = "^7.4"

[tool.ruff]
line-length = 120
indent-width = 2
exclude = ["tests", "app/templates"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
exclude = ["\\.venv"]

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"
